refactor: 패키지명 명명규칙에 맞게 수정
