+++ README.md
@@ -71,6 +71,7 @@
     - [x] 상속불가 클래스에 final 추가
     - [x] 정적 팩토리 메서드 네이밍 컨벤션에 맞게 수정
 - [x] PLAYER_NAME_DELIMITER 제거 및 로직 변경
+- [x] 패키지 명명규칙에 맞게 이름 수정
 - [ ] 패키지에 위치한 클래스들 하위에 패키지를 더 나누기
 - [ ] Table 클래스 이름 변경, 클래스 역할 고민
 - [ ]
