+++ src/main/java/blackjack/domain/card/cardelement/Suit.java
@@ -1,4 +1,4 @@
-package blackjack.domain.card.cardElement;
+package blackjack.domain.card.cardelement;
 
 public enum Suit {
     SPADE("스페이드"),
