+++ src/main/java/rentalcompany/car/Avante.java
@@ -1,11 +1,11 @@
-package rentCompany.car;
+package rentalcompany.car;
 
-public class Sonata extends Car {
-    private static final int DISTANCE_PER_LITER = 10;
-    private static final String NAME = "Sonata";
+public final class Avante extends Car {
+    private static final int DISTANCE_PER_LITER = 15;
+    private static final String NAME = "Avante";
     private final int tripDistance;
     
-    public Sonata(final int distance) {
+    public Avante(final int distance) {
         this.tripDistance = distance;
     }
