+++ src/test/java/blackjack/domain/card/DenominationTest.java
@@ -2,7 +2,7 @@ package blackjack.domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.domain.card.cardElement.Denomination;
+import blackjack.domain.card.cardelement.Denomination;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
