+++ src/test/java/blackjack/domain/human/PlayerTest.java
@@ -3,8 +3,8 @@ package blackjack.domain.human;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.card.Card;
-import blackjack.domain.card.cardElement.Denomination;
-import blackjack.domain.card.cardElement.Suit;
+import blackjack.domain.card.cardelement.Denomination;
+import blackjack.domain.card.cardelement.Suit;
 import org.junit.jupiter.api.Test;
 
 @SuppressWarnings("NonAsciiCharacters")
