+++ src/test/java/rentalcompany/RentalCompanyTest.java
@@ -1,18 +1,18 @@
-package rentCompany;
+package rentalcompany;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
-import rentCompany.car.Avante;
-import rentCompany.car.K5;
-import rentCompany.car.Sonata;
+import rentalcompany.car.Avante;
+import rentalcompany.car.K5;
+import rentalcompany.car.Sonata;
 
-public class RentCompanyTest {
+public class RentalCompanyTest {
     private static final String NEWLINE = System.getProperty("line.separator");
     
     @Test
     public void report() {
-        RentCompany company = RentCompany.create(); // factory method를 사용해 생성
+        RentalCompany company = RentalCompany.create(); // factory method를 사용해 생성
         company.addCar(new Sonata(150));
         company.addCar(new K5(260));
         company.addCar(new Sonata(120));
